Refresh token using ASP cookies (#145)
* Initial MFA logic for Exelon companies

Code that enables SMS and Email verification, tested with one of Exelon's subsidiaries

* Update exelon.py based on feedback

Merged some more similar pieces of code together and separated out Text vs Phone as Call will likely never be an option we will support.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Handle site maintenance 

Coincidentally site went down during testing and revealed that it raises a different kind of error than invalid auth. This handles that logic since we rely on the redirect.

* Add support for two forms of MFA

Thanks to trickv's diagnoses we know there are two forms of MFA in Exelon's codebase. I was able to locate an account that does not have an associated phone number and verify the logic flow between what I am calling "forced" and "opted" MFA. Users who do not provide a phone are forced into MFA (AFAIK there is no non-MFA version anymore). Users who add a phone, will forever be stuck as opted in. This logic adds email verification through both flows.

* Support refreshing tokens using ASP cookies

First pass to try to store the ASP session and cookies which can be used to acquire a new token.

* Raise invalid auth on reauthentication

If token returns nothing, a reauthentication is needed because the original authentication flow broke. We have to raise an error to support this flow.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>