Swap behavior out for exelon mobile apps. (#151)
* Swap behavior out for exelon mobile apps.

* Improve error handling for exelon MFA.

We need to support legacy clients being upgraded and ensure that the MFA challenge is presented when there is a problem with the refresh token. The only system would provide an invalid auth, but not trigger the MFA. This was tested by corrupting login_data with a bad token.

* Improve readability based on reviewer feedback.

Minor adjustments to the style to make it easier for others to follow based on the comments in the pull request.